requirements.txt
Django
django-rest-framework
flake8
psycopg2
drf-spectacular


